pyproject.toml
[tool.poetry]
name = "agenda-back"
version = "0.1.0"
description = ""
authors = ["Horacio <horacio.chiarella@gmail.com>"]
readme = "README.md"
packages = [{include = "agenda_back"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
pydantic = "^2.1.1"
sqlalchemy = "^2.0.19"
pydantic-settings = "^2.0.2"
httpx = "^0.24.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
email-validator = "^2.0.0.post2"

[tool.poetry.group.dev.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
ruff = "^0.0.280"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
passlib = "^1.7.4"

[tool.poetry.group.test.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
pytest = "^7.4.0"
pytest-custom-exit-code = "^0.3.0"
pytest-cov = "^4.1.0"
ruff = "^0.0.280"
httpx = "^0.24.1"
sqlalchemy = "^2.0.19"
passlib = "^1.7.4"

[tool.poetry.group.prod.dependencies]
python = "^3.10"
fastapi = "^0.100.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
ignore = [
  "BLE001",
  "ANN101", # write self type when used as parameter in a method
  "ANN102", # write class type when used as parameter in a method
  "COM812",
  "RET504",
  "B008",
  "E266",
  "A003"
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  "__init__.py",
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv"
]

line-length = 79

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py310"

fix = true

[tool.ruff.pydocstyle]
convention = "pep257"

[tool.ruff.flake8-import-conventions.aliases]
numpy = "np"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 7

[tool.ruff.per-file-ignores]
"tests/*"=["S101", "PT022", "EXE002", "D103"]
"agenda_back/schemas/v1/validator.py"=[
  "TRY003",  # Checks for long exception messages that are not defined in the exception class itself.
  "EM101",   # Checks for the use of string literals in exception constructors.
  "PLR2004", # Checks for the use of unnamed numerical constants ("magic") values in comparisons.
  "D102",    # Checks for undocumented public method definitions.
  # "PLR1714", # Checks for mergable comparisons
]
